refactor: extract worker.rs
